Move remaining handlers to their own files
