+++ src/eventDispatchers/touchEventHandlers/onImageRendered.js
@@ -1,0 +1,22 @@
+import { getters } from './../../store/index.js';
+
+export default function (evt) {
+  const eventData = evt.detail;
+  const element = eventData.element;
+
+  const toolsToRender = getters.
+    touchTools().
+    filter(
+      (tool) =>
+        tool.element === element &&
+        (tool.mode === 'active' ||
+          tool.mode === 'passive' ||
+          tool.mode === 'enabled')
+    );
+
+  toolsToRender.forEach((tool) => {
+    if (tool.renderToolData) {
+      tool.renderToolData(evt);
+    }
+  });
+}
